made a decent games page! and updated commission page too.
it's not perfect but I composed the games page the best I could with the work I had weh. not too shabby, could update one day with a clearer brain lol.

Also removed the "character design" page in favor of having a dedicated "games" and a "3D" page. Felt the most natural with the work I have.

Then added JJ's drawing to the commission page, rearranged juuust a little.